Added CRUD functionality to edit and delete accounts, added input validation for user account creation

+++ src/components/DeleteAccount.jsx
@@ -25,42 +25,43 @@ function DeleteAccount() {
 
   const handleDoNotDelete=(event)=>{
     event.preventDefault()
-    const user = username
-    fetch(`http://localhost:8080/user/${userID}/update`, {
-      method:"PUT",
-      headers:{"Content-Type":"application/json"},
-      body:JSON.stringify(user)
-    }).then(()=>{
-        alert("We're happy you stayed!")
-    }).then(event =>  window.location.href=`/myaccount/${username}`) // Redirects back to user's profile
+    alert("We're happy you stayed!");
+    window.location.href=`/myaccount/${username}`;
     }
 
 
   const handleDelete=(event)=>{
     event.preventDefault()
     const user = username
-    fetch(`http://localhost:8080/user/${userID}/delete`, {
-      method:"DELETE",
-      headers:{"Content-Type":"application/json"},
-      body:JSON.stringify(user)
-    }).then(()=>{
-        alert("Account Deleted!")
-    }).then(event =>  window.location.href=`/create`) // Redirects back to user's profile
-    }
+    if (confirm("Are you SURE you want to delete your account?!") == true) {
+      fetch(`http://localhost:8080/user/${userID}/delete`, {
+        method:"DELETE",
+        headers:{"Content-Type":"application/json"},
+        body:JSON.stringify(user)
+      }).then(()=>{
+          alert("Account Deleted!")
+      }).then(event =>  window.location.href=`/create`) // Redirects back to user's profile
+    } else alert("We're happy you stayed!")
+  }
 
   return (
-    <div>
-      <h1>Are you REALLY REALLY sure you want to delete {username}'s Account?</h1>
-      <p>This action cannot be undone!</p>
-
-      <Button variant="contained" onClick={event =>  window.location.href=`/myaccount/${username}`}>
-        No, wait! I'm having second thoughts! Get me outta here!
-      </Button >
-
-      <Button variant="contained" onClick={handleDelete}>
-        Yes, I'm sure, delete my account!
-      </Button >
-
+    <div style={{border: '5px solid rgba(0, 0, 0, 0.96)', padding: '50px', borderRadius: '25px'}}>
+      <div>
+        <h1>Are you<br></br>REALLY REALLY<br></br>sure you want to delete<br></br>{username}'s<br></br>Account?</h1>
+      </div>
+      <div>
+        <p>This action cannot be undone!</p>
+      </div>
+      <div>
+        <Button variant="contained" onClick={handleDoNotDelete} style={{margin : '5px'}}>
+          No, wait! I'm having second thoughts! Get me outta here!
+        </Button >
+      </div>
+      <div>
+        <Button variant="contained" onClick={handleDelete} style={{margin : '5px', backgroundColor: "red"}}>
+          Yes, I'm sure, delete my account!
+        </Button >
+      </div>
     </div>
   );
